[RELEASE] on release deploy to deploy lra-coordinator-quarkus-runner.jar as well
When we do the release we use the -Prelease which activates the release profile.
The maven behaviour for <activeByDefault>true</activeByDefault> is that activeByDefault profile is not activated
when another profile is activated explicitly with `mvn -P....`.
As we use -Prelease for release then the uber-jar profile is not activated and the -runnner.jar is not deployed to Nexus.
The uber jar profile has to be activated regardless if -P is or isn't used.